+++ src/org/maia/swing/animate/overlay/ColorOverlayComponent.java
@@ -322,7 +322,7 @@ public class ColorOverlayComponent extends BaseAnimatedComponent {
 		}
 
 		@Override
-		protected void updateStateBetweenPaints(Graphics2D g, double elapsedTimeMillis) {
+		protected void updateStateBetweenPaints(Graphics2D g, long elapsedTimeMillis) {
 			setPreviousTranslucency(getTranslucency());
 			updateStateOverTime();
 		}
